% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond.cov.R
\name{cond.cov}
\alias{cond.cov}
\title{Conditional covariance estimation by kernel-smoothing}
\usage{
cond.cov(datax, datay, x0, h, condmean = NULL)
}
\arguments{
\item{datax}{vector or matrix of observations of the conditioning variable}

\item{datay}{vector or matrix of observations of the conditioned variable,
with the same number of observations as \code{datax}}

\item{x0}{point at which the conditional covariance matrix is estimated.}

\item{condmean}{known or estimated conditional mean of \eqn{(Y_1,..., Y_d)}}
}
\description{
Computed an estimator of the
conditional covariance matrix of Y given X = x
}
\examples{
n = 10
X <- 1:n
mu_1 <- X^2
mu_2 <- 3*X
sd_1 <- X
sd_2 <- 1 + X
cor <- sin(X / 20)

# Computation of the covariance matrices
cov <- cor * sd_1 * sd_2
Sigma = array(dim = c(2,2,n))
Sigma[1,1,] <- sd_1^2
Sigma[2,2,] <- sd_2^2
Sigma[1,2,] <- Sigma[2,1,] <- cov

SqrtSigma = apply(X = Sigma, MARGIN = c(3), FUN = chol)
SqrtSigma = array(SqrtSigma, dim = c(2,2,n))
Z = matrix(rnorm(n = 2*n), ncol = 2)

Y1 = mu_1 + Z[,1] * SqrtSigma[1,1,] + Z[,2] * SqrtSigma[1,2,]
Y2 = mu_2 + Z[,1] * SqrtSigma[2,1,] + Z[,2] * SqrtSigma[2,2,]
datay = cbind(Y1, Y2)

# Comparison with the usual (unconditional) covariance
condcov = cond.cov(datax = X, datay,
                   x0 = 4, h = 1000)
print(condcov)
print(cov(datay))

# We now do smoothing
condcov = cond.cov(datax = X, datay,
                   x0 = 4, h = 2)
print(condcov)
print(cov(datay))

}
