% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condmoments.R
\name{cond.kurtosis}
\alias{cond.kurtosis}
\title{Conditional kurtosis estimation by kernel smoothing}
\usage{
cond.kurtosis(datax, datay, x0, h, condmean = NULL)
}
\arguments{
\item{datax}{the vector of observed X values.}

\item{datay}{the vector of observed Y values. Must have the same length as \code{datax}}

\item{x0}{the point at which the U-statistic should be computed.}

\item{h}{the bandwidth of the kernel smoothing.}

\item{condmean}{value of the conditional mean (if know or estimated previously).}
}
\description{
Conditional kurtosis estimation by kernel smoothing
}
\examples{
n = 500
datax = rnorm(n)
datay = datax + rnorm(n)
x0 = 1
h = 0.2
cond.kurtosis(datax = datax, datay = datay, x0 = x0, h = h)

}
