% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condUStat.R
\name{cond.UStat.univ}
\alias{cond.UStat.univ}
\title{Computation of a univariate conditional U-statistics by kernel smoothing}
\usage{
cond.UStat.univ(datax, datay, FUN, p, x0, h, cutoff = 8)
}
\arguments{
\item{datax}{the vector of observed X values.}

\item{datay}{the vector of observed Y values. Must have the same length as \code{datax}}

\item{FUN}{the kernel of the U statistic.
Should be vectorized and take as parameters \code{y1}, ..., \code{yp}.
(ie \code{p} vectors of the same length)}

\item{p}{the degree of the U-statistic given by FUN.}

\item{x0}{the point at which the U-statistic should be computed.}

\item{h}{the bandwidth of the kernel smoothing.}

\item{cutoff}{value above which the centered observations (X - x0)/h
have no influence.}
}
\description{
Computation of a univariate conditional U-statistics by kernel smoothing
}
\examples{
n = 500
datax = rnorm(n)
datay = datax + rnorm(n)
x0 = 1
h = 0.2
# Estimation of a conditional mean
cond.UStat(datax, datay, FUN = function(y1){return(y1)}, p=1, x0, h)
# Naive estimation of a conditional variance
cond.UStat(
  datax = datax, datay = datay,
  FUN = function(y1, y2){return(y1^2 - y1 * y2)}, p=2,
  x0 = c(x0,x0), h = h)

}
