% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condmoments.R
\name{cond.var}
\alias{cond.var}
\title{Conditional variance estimation by kernel smoothing}
\usage{
cond.var(datax, datay, x0, h, condmean = NULL, cutoff = 8)
}
\arguments{
\item{datax}{the vector of observed X values (conditioning variable).}

\item{datay}{the vector of observed Y values (conditioned variable).
Must have the same length as \code{datax}}

\item{x0}{the point at which the U-statistic should be computed.}

\item{h}{the bandwidth of the kernel smoothing.}

\item{condmean}{value of the conditional mean of y at x=x0
(if known or estimated previously).}
}
\description{
Conditional variance estimation by kernel smoothing
}
\examples{
n = 500
datax = rnorm(n)
datay = datax + rnorm(n)
x0 = 1
h = 0.2
cond.var(datax = datax, datay = datay, x0 = x0, h = h)

}
